Move LP Token Balance to separate component
Move LP Token Balance section to a separate component
