+++ solidity/dashboard/src/components/liquidity/index.js
@@ -1,4 +1,5 @@
 import APY from "./APY"
 import ShareOfPool from "./ShareOfPool"
+import LPTokenBalance from "./LPTokenBalance"
 
-export { APY, ShareOfPool }
+export { APY, ShareOfPool, LPTokenBalance }
